+++ CHANGELOG.md
@@ -171,7 +171,7 @@
 
 ## [1.3.1] - 2017-02-26
 * Add readme & changelog
-* Update gulpfile.js to enable theme packaging
+* Update Gulpfile.js to enable theme packaging
 * Remove font variables in functions.php
 
 ## [1.3.0] - 2017-02-26